Separate platform-specific layout from shared logic
