--- src/view/com/home/FeedsTabBar.tsx
@@ -1,1 +1,0 @@
-export * from './FeedsTabBarMobile'
