+++ src/view/com/home/HomeHeaderLayout.tsx
@@ -1,0 +1,1 @@
+export {HomeHeaderLayoutMobile as HomeHeaderLayout} from './HomeHeaderLayoutMobile'
